pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "derim-middleware"
version = "0.1.0"
description = "Smart Grid Digital Twin Middleware for Distributed Energy Resource Integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "DERIM Contributors"},
]
keywords = [
    "smart-grid",
    "digital-twin",
    "distributed-energy-resources",
    "DER",
    "middleware",
    "IEEE-2030.5",
    "CIM",
    "modbus",
    "mqtt",
    "ocpp",
    "sunspec",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "httpx>=0.25.0",
    "pymodbus>=3.6.0",
    "paho-mqtt>=1.6.0,<2.0.0",
    "ocpp>=0.26.0",
    "websockets>=12.0",
    "influxdb-client>=1.38.0",
    "aiosqlite>=0.19.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.3.0",
    "torch>=2.1.0",
    "matplotlib>=3.8.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.6.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/iceccarelli/DERIM-Middleware-project"
Documentation = "https://github.com/iceccarelli/DERIM-Middleware-project/tree/main/docs"
Repository = "https://github.com/iceccarelli/DERIM-Middleware-project"
Issues = "https://github.com/iceccarelli/DERIM-Middleware-project/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
known_first_party = ["derim"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/derim --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
